Set up terminal before fork.
After the fork it is undefined wether parent or child runs
first. So there can be a race: if the child runs before the
terminal of the parent is set up correctly luit may hang.
This patch sets up the terminal before forking and undoes
the settings in the child.

Signed-off-by: Mike FABIAN <mfabian@redhat.com>
Signed-off-by: Egbert Eich <eich@freedesktop.org>