Fix LNEXT handling on BSD
http://marc.info/?l=openbsd-misc&m=145123363601391&w=2

We need to turn off IEXTEN to make the tty transparent.  I thought
ICANON would cover this (as it does on Linux), but apparently it's a
BSD-ism that LNEXT operates independently of ICANON.  That isn't a bug
either; both the OpenBSD code and the rather different FreeBSD code
explicitly handle it this way.

IEXTEN is part of POSIX, even if the actual behavior is
"implementation-defined", so I don't think it needs an ifdef.

Signed-off-by: Matthieu Herrb <matthieu@herrb.eu>